rename utilComponent to ProfileButton
