+++ src/app/views/authentication/auth-util-components/index.tsx
@@ -1,1 +1,1 @@
-export { showSignInButtonOrProfile } from './UtilComponents';
+export { showSignInButtonOrProfile } from './ProfileButton';
